new runflags format: R123, S3+
The first letter determines (combination of) flags, the rest
sets runlevels. No default runlevels, except when no runlevels
are specified, in which case it's all-but-R0.

A dedicated leading letter inverts runlevel mask:
	R123	C_ONCE, levels -123------
	X123	C_ONCE, levels 0---456789
This is only available for R, because that's likely the only
combination that need inverted mask atm. If necessary, more
options may be added later.

This must be 4th or 5th change in inittab format.
I'm fucking crazy and it fucking shows.