configure MIPS qemu image to use serial console
Current configuration starts with no graphic output,
which makes the system unusable. The kernel already uses ttyS0 for output.

To make this all work out of the box on arches with and without graphics,
let's put both tty0 and ttyS0 gettys in inittab, and drop one of them
in post_build.sh. This way MIPS will be usable, and x86/ARM will be able
to use ttyS0 for gdb right away.