useful default runlevels mask
With rare exceptions, all entries should have 34...9 bits set
in their masks. Let's make it the default, so that ugly 3+ on
every single entry would not be necessary.

The patch also revokes "secondary recovery" meaning for level 2.
I can't think of any good use case that would require it to be
left out of the default mask.